Fix shared clipboard support
Installer tries to create /usr/bin/VBoxClient-all as a symlink to
${INSTALL_DIR}/other/98vboxadd-xclient, and has vboxclient.desktop use
that symlink for its Exec. Creation of the symlink fails, and we don't
want it in /usr/bin anyway, so just edit the .desktop file to point
directly to the script. This script launches the various VBoxClient
daemons to handle shared clipboard, drag and drop, display management,
and seamless window services.

We have to modify both copies because a startup script copies the
version from ${INSTALL_DIR} into the appropriate desktop manager paths
each boot.